Extract getBook and formatData from SearchViaISBN
